publish add obdiag-mcp
